refactor: remove test home page
